refactor(copilot): Extract the hide-debounce and ref-count the section observable
Pull the show-immediate / hide-after-delay operator out of the
UaiCopilotContext constructor into a standalone debouncedHide() so the
supported-workspace signal can be unit tested without constructing the
whole context. Behaviour is unchanged.

Also switch the shared section observable to shareReplay({ refCount: true })
so its history-patch teardown actually runs when the last subscriber
leaves (and re-applies on re-subscribe). Under the previous shareReplay(1)
the source stayed subscribed for the app lifetime, so that teardown was
dead code. No user-facing impact today (the observable has one
app-lifetime subscriber), but it makes the ref-count design real and
testable.
